requirements.txt
amqp==5.0.5
asgiref==3.3.1
autopep8==1.5.5
beautifulsoup4==4.9.3
billiard==3.6.3.0
blessings==1.7
bpython==0.21
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
curtsies==0.3.5
cwcwidth==0.1.4
Django==3.1.6
djangorestframework==3.12.2
greenlet==1.0.0
idna==2.10
kombu==5.0.2
Markdown==3.3.3
model-bakery==1.2.1
prompt-toolkit==3.0.16
pycodestyle==2.6.0
Pygments==2.8.0
pytz==2021.1
pyxdg==0.27
redis==3.5.3
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
